feat(bootstrap): #1587 refactor bootstrap wiring + config loader split
Split monolithic bootstrap files into domain-specific subdirectories:
- factory/config/ (config_loader, config_models, config_deprecation)
- factory/hub/ (hub_core, hub_assembly, hub_agent_registration, hub_cli_pool,
  hub_llm_client, hub_inbound_bus, hub_clipool_init)
- wiring/auth.py + wiring/ingest_wiring.py extracted from bootstrap_wiring.py
- factory/providers.py for provider registry

Backward-compatible imports maintained via barrel __init__.py files.
Tests patched to target new module locations.

Quality gates:
- 5286 tests pass
- ruff / pyright / import-linter / file-length / folder-size / duplicate-test-
  basenames / no-runtime-toml-bots all pass
- file_exemptions.txt: removed 3 entries (config.py deleted, hub_builder.py
  and agent_factory.py shrunk below cap)
