+++ src/lyra/bootstrap/factory/agent_factory.py
@@ -10,7 +10,8 @@ from typing import TYPE_CHECKING
 # Re-exported for backward compatibility (tests import these from agent_factory)
 from lyra.agents.simple_agent import SimpleAgent
 from lyra.bootstrap.factory.bot_agent_map import (
-    resolve_bot_agent_map,  # noqa: F401 — DEBT:re-export-init
+    _resolve_bot_agent_map,
+    resolve_bot_agent_map,  # noqa: F401 — DEBT:re-export-init  # type: ignore
 )
 from lyra.bootstrap.factory.config import (
     LlmConfig,
@@ -18,8 +19,12 @@ from lyra.bootstrap.factory.config import (
     _load_circuit_config,
     _load_messages,
 )
+from lyra.bootstrap.factory.providers import (
+    _build_per_agent_registry,
+    _build_shared_base_providers,
+)
 from lyra.bootstrap.types import BotAuthBundle
-from lyra.bootstrap.wiring.bootstrap_wiring import BotAuthDeps, _build_bot_auths
+from lyra.bootstrap.wiring.auth import BotAuthDeps, _build_bot_auths
 from lyra.config import multibot_config_from_store
 from lyra.core.agent import Agent, AgentBase
 from lyra.core.agent.agent_loader import agent_row_to_config
@@ -32,8 +37,6 @@ from lyra.infrastructure.stores.agent_store import AgentStore
 from lyra.integrations.base import SessionTools
 from lyra.integrations.vault_cli import VaultCli
 from lyra.integrations.web_intel import WebIntelScraper
-from lyra.llm.base import LlmProvider
-from lyra.llm.decorators import CircuitBreakerDecorator, RetryDecorator
 from lyra.llm.drivers.cli import ClaudeCliDriver
 from lyra.llm.registry import ProviderRegistry
 
@@ -77,20 +80,6 @@ class ResolveAgentsDeps:
     llm_cfg: LlmConfig | None = None
     nats_llm_client: "LlmClient | None" = None
     cli_nats_driver: "LlmClient | None" = None
-
-
-# ---------------------------------------------------------------------------
-# Backward-compatible alias used by multibot.py
-# ---------------------------------------------------------------------------
-
-
-async def _resolve_bot_agent_map(
-    agent_store: AgentStore,
-    tg_bots: list,
-    dc_bots: list,
-) -> dict:
-    """Thin alias — delegates to bot_agent_map.resolve_bot_agent_map."""
-    return await resolve_bot_agent_map(agent_store, tg_bots, dc_bots)
 
 
 async def _init_bot_auths_and_agents(
@@ -160,89 +149,6 @@ async def _init_bot_auths_and_agents(
         circuit_registry=circuit_registry,
         admin_user_ids=admin_user_ids,
     )
-
-
-def _build_shared_base_providers(  # noqa: PLR0913
-    circuit_registry: CircuitRegistry,
-    cli_pool: CliPool | None,
-    llm_cfg: LlmConfig,
-    *,
-    nats_llm_client: "LlmClient | None" = None,
-    cli_nats_driver: "LlmClient | None" = None,
-    cb_decorator_cls: type = CircuitBreakerDecorator,
-    retry_decorator_cls: type = RetryDecorator,
-) -> dict[str, LlmProvider]:
-    """Build ``{backend: base LlmProvider}`` reusable across all agents.
-
-    ``claude-cli`` (ClaudeCliDriver or LlmClient via clipool), ``nats`` (Retry ->
-    LlmClient, only when ``nats_llm_client`` is provided). Callers layer
-    decorators per agent via ``_build_per_agent_registry``.
-
-    ``cli_nats_driver`` takes precedence over ``cli_pool`` for the
-    ``claude-cli`` backend when both are provided.
-    """
-    providers: dict[str, LlmProvider] = {}
-
-    if cli_nats_driver is not None:
-        cli_cb = circuit_registry.get("claude-cli")
-        base: LlmProvider = cli_nats_driver
-        if cli_cb is not None:
-            providers["claude-cli"] = cb_decorator_cls(base, cli_cb)
-        else:
-            providers["claude-cli"] = base
-        log.info("Shared base: built claude-cli driver via NATS (decorated)")
-    elif cli_pool is not None:
-        cli_driver: LlmProvider = ClaudeCliDriver(cli_pool)
-        cli_cb = circuit_registry.get("claude-cli")
-        if cli_cb is not None:
-            providers["claude-cli"] = cb_decorator_cls(cli_driver, cli_cb)
-        else:
-            providers["claude-cli"] = cli_driver
-        log.info("Shared base: built claude-cli driver (in-process, decorated)")
-
-    if nats_llm_client is not None:
-        providers["nats"] = retry_decorator_cls(
-            nats_llm_client,
-            max_retries=llm_cfg.max_retries,
-            backoff_base=llm_cfg.backoff_base,
-        )
-        log.info("Shared base: registered nats driver (decorated)")
-
-    return providers
-
-
-def _build_per_agent_registry(
-    shared_providers: dict[str, LlmProvider],
-) -> ProviderRegistry:
-    """Build a per-agent ProviderRegistry on top of shared driver instances.
-
-    ``shared_providers`` is the dict returned by ``_build_shared_base_providers``.
-    For each backend, the provider is registered as-is.
-    """
-    registry = ProviderRegistry()
-
-    for backend, base_provider in shared_providers.items():
-        registry.register(backend, base_provider)
-
-    return registry
-
-
-def _build_provider_registry(
-    circuit_registry: CircuitRegistry,
-    cli_pool: CliPool | None,
-    llm_cfg: LlmConfig | None = None,  # None -> LlmConfig() (defaults)
-) -> ProviderRegistry:
-    """Build and return a ProviderRegistry with all configured drivers.
-
-    Convenience wrapper used by the legacy single-agent bootstrap path.
-    For multi-agent startup use ``_build_shared_base_providers`` +
-    ``_build_per_agent_registry`` to avoid rebuilding the driver stack per
-    agent.
-    """
-    shared = _build_shared_base_providers(
-        circuit_registry, cli_pool, llm_cfg or LlmConfig(), cli_nats_driver=None
-    )
-    return _build_per_agent_registry(shared)
 
 
 def _create_agent(deps: CreateAgentDeps) -> AgentBase:
