+++ src/lyra/bootstrap/factory/hub/__init__.py
@@ -1,0 +1,1 @@
+"""Hub construction domain helpers."""
