+++ tools/file_exemptions.txt
@@ -1,5 +1,4 @@
 # File size exemptions (≤300 lines per file)
-src/lyra/bootstrap/factory/config.py  # 353 lines — DEBT:file-exemptions — #1467 +23 lines (AdapterConfigBundle dataclass + build_adapter_config_bundle factory); #1420 +43 lines (warn_deprecated_bot_sections + _bot_sections_deprecation_warned flag + call site); config loader aggregator — all _load_* helpers live here
 # Each entry must have a tracking issue.
 # Format: <path>  # <lines> lines — <issue> <description>
 tests/core/conftest.py  # 314 lines — DEBT:file-exemptions — backward-compatible re-exports after factory extraction (#1417)
@@ -22,6 +21,4 @@ src/lyra/adapters/discord/discord_inbound.py  # 346 lines — DEBT:file-exemptio
 src/lyra/core/auth/authenticator.py  # 337 lines — DEBT:file-exemptions — #1494 AuthenticatorDeps + FromBotStoreDeps frozen DI dataclasses (wiring-bootstrap-deps paydown — __init__ and from_bot_store collapsed off PLR0913)
 src/lyra/core/commands/command_router.py  # 321 lines — DEBT:file-exemptions — #1494 +11 lines (CommandRouterDeps frozen DI dataclass — wiring-bootstrap-deps paydown);#1434 base_dir param + _base_dir wiring for admin path-traversal fix
 src/lyra/monitoring/checks.py  # 324 lines — DEBT:file-exemptions — #1068 +6 lines (blobstore disk_pct + inode_pct checks); #1482 T11 +18 lines (audio consumer lag + stream fullness checks wired into run_checks)
-src/lyra/bootstrap/factory/agent_factory.py  # 351 lines — DEBT:file-exemptions — #1283 Phase 6: _init_bot_auths_and_agents + _resolve_bot_agent_map + _create_agent + _build_provider_registry extracted into agent_factory.py (bootstrap DI container migration); #1492 -4 lines (dropped redundant ValueError catch-and-reraise, traceback now preserved)
-src/lyra/bootstrap/factory/hub_builder.py  # 370 lines — DEBT:file-exemptions — #1494 +3 lines (deps-object call-site migration — wiring-bootstrap-deps paydown);#1283 Phase 6: _build_hub + _init_clipool + _build_hub_and_wire extracted into hub_builder.py (bootstrap DI container migration); pre-existing build_inbound_bus + build_llm_client; #1492 -53 lines (deleted dead public build_hub parallel path + dead audit_sink provisioning + dead inbound_bus_cfg param); #1540 +5 lines (init_blobstore import + blob_store on BuildHubDeps — ADR-082)
 src/lyra/core/tts_dispatch.py  # 305 lines — DEBT:file-exemptions — #1553 +6 lines (None-guard: SynthesisResult.blob_ref became Optional after PENDING_STORE_KEY retirement; narrows BlobRef|None→BlobRef before OutboundAudio construction + logs unexpected-None). File was 299; consider absorbing when AudioPipeline next splits.
